requirements.txt
streamlit==1.32.0
langgraph==0.6.6
langchain==0.3.27
langchain-openai==0.3.32
langchain-tavily==0.2.11
langchain-community==0.3.29
python-dotenv==1.0.0
fastapi==0.109.0
uvicorn==0.27.0
requests==2.32.5
beautifulsoup4==4.12.2
pydantic==2.11.7
